camera offset and gravity

+++ src/comps/offset.h
@@ -7,6 +7,6 @@ namespace Comps
 	// Stores an offset in a Vect<int32_t> object
 	struct Offset
 	{
-		Vect<int32_t> offset;
+		Vect<float> offset;
 	};
 }
